Disable tests or test cases that access singularity hub
While singulariy hub is down, disable the tests that try to access it.
Every disabled test or test case has been marked with a comment
indicating the reason for disabling it is the fact that singularity hub
is down. This is done in order to avoid confusion in case the code is
reworked before the tests can be reenabled again.

This commit applies https://github.com/sylabs/singularity/pull/3865/commits/0ce7e0b61a1b7bcb0d79a1dbaec795c9ca230111
manually in order to prevent the unnecessary creation and changing of files
that are not within this branch of singularity

Signed-off-by: Ian Kaneshiro <iankane@umich.edu>