Do not set blkio cgroup limits during test
blkio requires an IO scheduler, and there's no gurantee one is available
when running tests (recent kernels disable the IO scheduler when the
block device is a virtual disk, e.g. vda, as it's the case with some
virtual machines). Trying to make this conditional on the existance of
an IO scheduler seems more complicated than what it's worth, as the
TestCgroups test is testing more whether it can load the configuration
file and less that everything in that file is applied correctly (if it
checks anything, it's only the stuff directly related to CPU). Most of
the logic that applies the configuration file is already tested as part
of the cgroups package anyways.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>