Basic CircleCI for pro-3.1
Ensure a sane config for CircleCI on the pro-3.1 branch so that we are at
least running the basic test suite here, even if it is not on the
distribuitons that we are targeting.

This drops MacOS and alpine build/check since we don't support that in Pro.
Go 1.13 is now used here, as RHEL is now on 1.13 from EPEL and we manually
setup 1.13 in the infra repository on other platforms.