Initial debian packaging of singularitypro35
This allows build of a singularitypro35 deb package via:

  dpkg-buildpackage -b -r<root_command> -us -uc

Where <root_command> is 'sudo' or 'fakeroot' or similar as configured on
your system.

The packaging here is *not* compliant with Debian upstream type things
at all. We are assuming that the required version of Go to build is
present and on PATH. It is *not* declared as a build dependency here.

The intent is to build packages using the VM instances configured by the
singularity-pro-infra terraform and ansible configuration.

You *cannot* build with pbuilder or sbuild etc.