Hack for pulling version & release from a git tag.
Pro needs to use non sem-ver version-release, e.g. 3.5-1

In this case, `3.5-1` we need:

   - package_version = 3.5-1 (for --version etc)
   - short_version   = 3.5   (for rpm version)
   - release_info    = 1     (for rpm release)

The existing split of a package_version into short_version and release
_info is based on a colon, but this is not a valid character in a git
tag. This means we cannot set a correct rpm version & release from a git
tag. We have to use the `-V` to mconfig manually. Using -V manually is
not nice from an automation perspective.

This PR works around it by allowing short_version and release_info to be
separated by '@'.

We also look for tags beginning `pro-` rather than `v` which are
specific to Singularity PRO releases, so we don't get confused from
merges of the OSS repo.

A tag `pro-3.5@1` will result in:

   - package_version = 3.5-1 (for --version etc)
   - short_version   = 3.5   (for rpm version)
   - release_info    = 1     (for rpm release)